Create entrypoint.go
